Did some correction
Hello, I have updated the format of the existing links that were present in the 'Resources' section as the right markdown syntax was not used and they were not visible, I have corrected the same and now it is visible. Also, added a new resource for CSS selectors - 'Universal Selector in CSS'. Hope this small contribution will help other learners.